[5.6] Delete code duplication in findOrFail and firstOrFail methods for relations.
