+++ src/Illuminate/Database/Eloquent/Relations/HasManyThrough.php
@@ -6,10 +6,14 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Database\Eloquent\Relations\Traits\FindOrFail;
+use Illuminate\Database\Eloquent\Relations\Traits\FirstOrFail;
 
 class HasManyThrough extends Relation
 {
+    use FirstOrFail,
+        FindOrFail;
+
     /**
      * The "through" parent model instance.
      *
@@ -260,23 +264,6 @@ class HasManyThrough extends Relation
     }
 
     /**
-     * Execute the query and get the first result or throw an exception.
-     *
-     * @param  array  $columns
-     * @return \Illuminate\Database\Eloquent\Model|static
-     *
-     * @throws \Illuminate\Database\Eloquent\ModelNotFoundException
-     */
-    public function firstOrFail($columns = ['*'])
-    {
-        if (! is_null($model = $this->first($columns))) {
-            return $model;
-        }
-
-        throw (new ModelNotFoundException)->setModel(get_class($this->related));
-    }
-
-    /**
      * Find a related model by its primary key.
      *
      * @param  mixed  $id
@@ -310,30 +297,6 @@ class HasManyThrough extends Relation
         return $this->whereIn(
             $this->getRelated()->getQualifiedKeyName(), $ids
         )->get($columns);
-    }
-
-    /**
-     * Find a related model by its primary key or throw an exception.
-     *
-     * @param  mixed  $id
-     * @param  array  $columns
-     * @return \Illuminate\Database\Eloquent\Model|\Illuminate\Database\Eloquent\Collection
-     *
-     * @throws \Illuminate\Database\Eloquent\ModelNotFoundException
-     */
-    public function findOrFail($id, $columns = ['*'])
-    {
-        $result = $this->find($id, $columns);
-
-        if (is_array($id)) {
-            if (count($result) == count(array_unique($id))) {
-                return $result;
-            }
-        } elseif (! is_null($result)) {
-            return $result;
-        }
-
-        throw (new ModelNotFoundException)->setModel(get_class($this->related));
     }
 
     /**
