+++ src/Illuminate/Database/Eloquent/Relations/Traits/FindOrFail.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace Illuminate\Database\Eloquent\Relations\Traits;
+
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+
+/**
+ * Trait FindOrFail.
+ *
+ * @property $related
+ */
+trait FindOrFail
+{
+    /**
+     * Find a related model by its primary key or throw an exception.
+     *
+     * @param  mixed  $id
+     * @param  array  $columns
+     * @return \Illuminate\Database\Eloquent\Model|\Illuminate\Database\Eloquent\Collection
+     *
+     * @throws \Illuminate\Database\Eloquent\ModelNotFoundException
+     */
+    public function findOrFail($id, $columns = ['*'])
+    {
+        $result = $this->find($id, $columns);
+
+        if (is_array($id)) {
+            if (count($result) == count(array_unique($id))) {
+                return $result;
+            }
+        } elseif (! is_null($result)) {
+            return $result;
+        }
+
+        throw (new ModelNotFoundException)->setModel(get_class($this->related));
+    }
+}
